rest/results-tabulation-api/requirements.txt
alembic==0.8.5
connexion==2.2.0
decorator==4.4.0
Flask==1.0.2
Flask-Caching==1.7.2
Flask-Cors==3.0.8
Flask-Migrate==1.8.0
flask-marshmallow==0.10.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.4.0
marshmallow==2.19.2
marshmallow-enum==1.4.1
marshmallow-sqlalchemy==0.16.3
num2words==0.5.10
numpy==1.18.0
openapi-spec-validator==0.2.7
pandas==0.25.3
pyopenssl==19.0.0
pytest==5.1.2
python-jose==3.0.1
psycopg2-binary==2.8.2
PyMySQL==0.9.3
SQLAlchemy==1.3.3
swagger-ui-bundle==0.0.3
